Prettify checkboxes and radio buttons
Use UniformJS to prettify form elements, such as checkboxes and radio
buttons. There were minor template workarounds to, unfortunately, workaround
some things that Uniform does not do correctly. For example: for a
checkbox that has a label associated with it, both the checkbox + label
should be wrapped within <label>..</label> for them to be functional.