Fix issues with GA
The event is submitted to Google asynchronously. When we do the form
submit, the context of the page switches and the event submission no
longer gets processed.

To overcome this, we push the trackEvent command to Google's Analytics
queue. Once that has been pushed to the queue, we push the command to
submit the form.