Workaround for GA async event tracking
Since pushing the event to GA's queue did not work, we use a small hack
to track events + submit the download form. We push the event to GA's
async queue and then submit the form 300 milliseconds later. This should
work (in theory) and should not result in a noticable user-facing change
at the same time.